fix: Ensure proper type casting for cache size comparisons in gamma, zeta, and beta functions

+++ maths.hpp
@@ -33,9 +33,9 @@ const BigInt &gamma(int64_t n) {
   if (n <= 0)
     throw std::domain_error("Gamma function is not defined for non-positive integers.");
   static std::vector<BigInt> cache{1_big, 1_big};
-  if (n <= cache.size())
+  if ((uint64_t)n <= cache.size())
     return cache[n - 1];
-  while (cache.size() < n)
+  while (cache.size() < (uint64_t)n)
     cache.emplace_back(cache.back() * BigInt((uintmax_t)cache.size()));
   return cache[n - 1];
 }
@@ -54,9 +54,9 @@ const Fraction &zeta(int64_t n) {
     throw not_implemented("Zeta function for odd n");
   n >>= 1; // n is now even
   static std::vector<Fraction> cache{Fraction(1)};
-  if (n < cache.size())
+  if ((uint64_t)n < cache.size())
     return cache[n];
-  for (size_t k = cache.size(); k <= n; ++k) {
+  for (size_t k = cache.size(); k <= (uint64_t)n; ++k) {
     cache.emplace_back(minus_one_power(k + 1));
     cache.back() *= Fraction((uintmax_t)k);
     cache.back() /= factorial(2 * k + 1);
@@ -84,9 +84,9 @@ const Fraction &beta(int64_t n) {
     throw not_implemented("Beta function for even n");
   n >>= 1; // Beta(2n + 1)
   static std::vector<Fraction> cache{Fraction(1, 4)};
-  if (n < cache.size())
+  if ((uint64_t)n < cache.size())
     return cache[n];
-  for (size_t k = cache.size(); k <= n; ++k) {
+  for (size_t k = cache.size(); k <= (uint64_t)n; ++k) {
     cache.emplace_back();
     for (size_t m = 1; m <= k; ++m) {
       switch (m & 1) {
